requirements.txt
aiohttp==3.8.6
aiokafka==0.8.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
confluent-kafka==2.3.0
elastic-transport==8.10.0
elasticsearch==8.11.0
exceptiongroup==1.1.3
fastapi==0.104.1
fastapi-cache2==0.2.1
frozenlist==1.4.0
h11==0.14.0
idna==3.4
kafka-python==2.0.2
multidict==6.0.4
packaging==23.2
pendulum==2.1.2
pydantic==2.5.1
pydantic_core==2.14.3
python-dateutil==2.8.2
pytzdata==2020.1
redis==5.0.1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
yarl==1.9.2
